stock/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.2
Django==3.1.14
django-cors-headers==3.7.0
django-filter==2.4.0
django-silk==4.2.0
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
gprof2dot==2019.11.30
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.9
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.21.2
openpyxl==3.0.9
packaging==20.8
pandas==1.3.3
priority==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==21.2.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.5
zope.interface==5.2.0
mysqlclient==2.1.0
django-apscheduler==0.4.2
APScheduler~=3.9.1
PyPDF2~=2.1.0
psycopg2~=2.9.3
tornado~=6.2
httpx~=0.23.0